Adds a quick spec to test the use case where paginative fields points to a column on a joint table
i.e. I have a query that joins my associated table, and I want to paginate based on a field on that joint table (let's say created at, as in the example), I need to setup my mapping to map `created_at` to the `created_at` field of that other table. This functionality was already introduced via the DSL setup, just needed some coverage.